pyproject.toml
[tool.poetry]
name = "karomag_gendiff"
version = "0.1.1"
description = ""
readme = "README.md"
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]
authors = ["Karomag <roma-novikov@mail.ru>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.13.3"
pytest = "^5.3.2"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/legacy/"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

